In tests, prefer %zu to %lu+cast
* tests/test-dirname.c (main):
* tests/test-gc-pbkdf2-sha1.c (main):
* tests/test-gc-pbkdf2.c (main):
* tests/test-nonblocking-reader.h (full_read)
* tests/test-nonblocking-writer.h (main_writer_loop):
(full_read_from_nonblocking_fd):
* tests/test-read-file.c (test_read_file):
Prefer %zu to %lu plus a cast to unsigned long.
* tests/test-nonblocking-reader.h (full_read)
* tests/test-nonblocking-writer.h (main_writer_loop):
(full_read_from_nonblocking_fd):
Pacify -Wuseless-cast via compound literals.

+++ tests/test-dirname.c
@@ -148,9 +148,9 @@ main (void)
         {
           ok = false;
           printf ("dir_name '%s': got '%s' len %d,"
-                  " expected '%s' len %lu\n",
+                  " expected '%s' len %zu\n",
                   t->name, dir, dirlen,
-                  t->dir, (unsigned long) strlen (t->dir));
+                  t->dir, strlen (t->dir));
         }
       if (!streq (last, t->last))
         {
@@ -165,9 +165,9 @@ main (void)
         {
           ok = false;
           printf ("base_name '%s': got '%s' len %d,"
-                  " expected '%s' len %lu\n",
+                  " expected '%s' len %zu\n",
                   t->name, base, baselen,
-                  t->base, (unsigned long) strlen (t->base));
+                  t->base, strlen (t->base));
         }
       if (!streq (stripped, t->stripped) || modified != t->modified)
         {
