+++ tests/test-read-file.c
@@ -59,8 +59,7 @@ test_read_file (int flags)
               /* FILE1 is a regular file or a symlink to a regular file.  */
               if (len != statbuf.st_size)
                 {
-                  fprintf (stderr, "Read %lu from %s...\n",
-                           (unsigned long) len, FILE1);
+                  fprintf (stderr, "Read %zu from %s...\n", len, FILE1);
                   err = 1;
                 }
             }
@@ -101,8 +100,7 @@ test_read_file (int flags)
              is not a regular file.  */
           if (len != 0)
             {
-              fprintf (stderr, "Read %lu from %s...\n",
-                       (unsigned long) len, FILE2);
+              fprintf (stderr, "Read %zu from %s...\n", len, FILE2);
               err = 1;
             }
           free (out);
